vehicleTrackingUpdate

+++ lib/src/src/models/vehicle.dart
@@ -7,7 +7,7 @@ class Vehicle {
   final double latitude;
   final double speed;
   final int deviceId;
-  final double km;
+  final int km;
   final bool isActive;
   final int sensors;
   final String plate;
